requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
psycopg2-binary==2.9.9
